requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-multiple-models==2.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tzdata==2021.5
whitenoise==5.3.0
